emacs: Integrate meow-mode and disable evil mode for now
gasp! It's true.  Evil mode has been a long standing companion for me
as I made my way from vim to nvim to emacs.  It's been great, it's a
faithul and well designed vim emulation layer on top of emacs, but it
_is_ an emulation layer.  It imposes foreign concepts on top of an
editor with pre-existing convetions, names, and editing interface.

Meow, by contrast embraces emacs for what it is, and simply adds a
modal layer on top of all the emacs editing goodness.  It also imports
kakoune/helix style selection->action editing grammar as opposed to
vim's action->selection grammar.