emacs: fix startup frame font settings
for the longest time, the initial emacs frame didn't have the same
font settings as what would ultimately be set by either making a new
frame or by changing the theme.

I've worked around this for a long time, but ultimately, the solution
was to add the font settings to the 'after-make-frame-functions hook
rather than the 'after-init-hook.